refactor: moved utils in util folder
