+++ __tests__/LottoTest.js
@@ -4,7 +4,7 @@ const {
   validateLottoBudget,
   validateBonusNumber,
   validateBonusNumberNotInLottoNumber,
-} = require('../src/validate');
+} = require('../src/utils/validate');
 
 describe('로또 클래스 테스트', () => {
   test('로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.', () => {
